requirements.txt
python-dotenv==1.0.1
pydantic==2.10.2
rich==13.9.4
typer==0.15.4
httpx==0.28.1
psutil>=5.9.0
pytest>=7.4.0
pytest-asyncio==0.23.0
